# atoma-daemon/docs/openapi.yml
openapi: 3.1.0
info:
  title: atoma-daemon
  description: ''
  license:
    name: Apache-2.0
  version: 0.1.0
servers:
- url: http://localhost:8080
  description: Local development server
paths:
  /almost_filled_stacks/{fraction}:
    get:
      tags:
      - almost_filled_stacks
      summary: Retrieves all stacks that are filled above a specified fraction threshold for all registered nodes.
      operationId: almost_filled_stacks_fraction_get
      parameters:
      - name: fraction
        in: path
        description: The fraction threshold (0.0 to 100.0) to filter stacks
        required: true
        schema:
          type: number
          format: double
      responses:
        '200':
          description: Stack objects that are filled above the specified fraction threshold
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Stack'
        '500':
          description: Internal server error
  /almost_filled_stacks/{id}/{fraction}:
    get:
      tags:
      - almost_filled_stacks
      summary: Retrieves all stacks that are filled above a specified fraction threshold for a specific node.
      description: |-
        # Arguments
        * `daemon_state` - The shared state containing the state manager
        * `node_small_id` - The small ID of the node whose stacks should be retrieved
        * `fraction` - The fraction threshold (0.0 to 100.0) to filter stacks

        # Returns
        * `Result<Json<Vec<Stack>>>` - A JSON response containing a list of stacks
          - `Ok(Json<Vec<Stack>>)` - Successfully retrieved stacks
          - `Err(StatusCode::INTERNAL_SERVER_ERROR)` - Failed to retrieve stacks from state manager

        # Example Response
        Returns a JSON array of Stack objects for the specified node that are filled above the fraction threshold
      operationId: almost_filled_stacks_get
      parameters:
      - name: id
        in: path
        description: The small ID of the node whose stacks should be retrieved
        required: true
        schema:
          type: integer
          format: int64
      - name: fraction
        in: path
        description: The fraction threshold (0.0 to 100.0) to filter stacks
        required: true
        schema:
          type: number
          format: double
      responses:
        '200':
          description: List of node stacks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Stack'
        '500':
          description: Internal server error
...
  /stacks/claimed_stacks:
    get:
      tags:
      - stacks
      summary: Retrieves all claimed stacks for the currently registered node badges.
      operationId: claimed_stacks_list
      responses:
        '200':
          description: List of all StackSettlementTicket objects for all registered nodes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/StackSettlementTicket'
        '500':
          description: Internal server error
  /stacks/claimed_stacks/{id}:
    get:
      tags:
      - stacks
      summary: Retrieves all claimed stacks for a specific node identified by its small ID.
      operationId: claimed_stacks_get
      parameters:
      - name: id
        in: path
        description: The small ID of the node whose claimed stacks should be retrieved
        required: true
        schema:
          type: integer
          format: int64
      responses:
        '200':
          description: List of StackSettlementTicket objects for the specified node
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/StackSettlementTicket'
        '500':
          description: Internal server error
  /stacks/stacks:
    get:
      tags:
      - stacks
      summary: Retrieves all stacks associated with the currently registered node badges.
      operationId: stacks_list
      responses:
        '200':
          description: List of all Stack objects for all registered nodes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Stack'
        '500':
          description: Internal server error
  /stacks/stacks/{id}:
    get:
      tags:
      - stacks
      summary: Retrieves all stacks for a specific node identified by its small ID.
      operationId: stacks_get
      parameters:
      - name: id
        in: path
        description: The small ID of the node whose stacks should be retrieved
        required: true
        schema:
          type: integer
          format: int64
      responses:
        '200':
          description: List of Stack objects for the specified node
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Stack'
        '500':
          description: Internal server error
  /subscriptions/:
    get:
      tags:
      - subscriptions
      summary: Retrieves all node subscriptions for the currently registered node badges.
      operationId: subscriptions_list
      responses:
        '200':
          description: List of all node subscriptions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/NodeSubscription'
        '404':
          description: No node badges registered
        '500':
          description: Internal server error
  /subscriptions/{id}:
    get:
      tags:
      - subscriptions
      summary: Retrieves all subscriptions for a specific node identified by its small ID.
      operationId: subscriptions_get
      parameters:
      - name: id
        in: path
        description: The small ID of the node whose subscriptions should be retrieved
        required: true
        schema:
          type: integer
          format: int64
      responses:
        '200':
          description: List of node subscriptions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/NodeSubscription'
        '500':
          description: Internal server error
  /tasks/:
    get:
      tags:
      - tasks
      summary: Retrieves all tasks from the state manager.
      operationId: tasks_list
      responses:
        '200':
          description: List of all tasks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Task'
        '500':
          description: Internal server error
components:
  schemas:
...
    NodeSubscription:
      type: object
      description: Represents a node subscription to a task
      required:
      - node_small_id
      - task_small_id
      - price_per_compute_unit
      - max_num_compute_units
      - valid
      properties:
        max_num_compute_units:
          type: integer
          format: int64
          description: Maximum number of compute units for the subscription
        node_small_id:
          type: integer
          format: int64
          description: Unique small integer identifier for the node subscription
        price_per_compute_unit:
          type: integer
          format: int64
          description: Price per compute unit for the subscription
        task_small_id:
          type: integer
          format: int64
          description: Unique small integer identifier for the task
        valid:
          type: boolean
          description: Indicates whether the subscription is valid
...
    Stack:
      type: object
      description: Represents a stack of compute units for a specific task
      required:
      - owner_address
      - stack_small_id
      - stack_id
      - task_small_id
      - selected_node_id
      - num_compute_units
      - price
      - already_computed_units
      - in_settle_period
      - total_hash
      - num_total_messages
      properties:
        already_computed_units:
          type: integer
          format: int64
          description: Number of compute units already processed
        in_settle_period:
          type: boolean
          description: Indicates whether the stack is currently in the settle period
        num_compute_units:
          type: integer
          format: int64
          description: Total number of compute units in this stack
        num_total_messages:
          type: integer
          format: int64
          description: Number of payload requests that were received by the node for this stack.
        owner_address:
          type: string
          description: Address of the owner of the stack
        price:
          type: integer
          format: int64
          description: Price of the stack (likely in smallest currency unit)
        selected_node_id:
          type: integer
          format: int64
          description: Identifier of the selected node for computation
        stack_id:
          type: string
          description: Unique string identifier for the stack
        stack_small_id:
          type: integer
          format: int64
          description: Unique small integer identifier for the stack
        task_small_id:
          type: integer
          format: int64
          description: Small integer identifier of the associated task
        total_hash:
          type: array
          items:
            type: integer
            format: int32
            minimum: 0
          description: |-
            Joint concatenation of Blake2b hashes of each payload and response pairs that was already processed
            by the node for this stack.
...
    StackSettlementTicket:
      type: object
      description: Represents a settlement ticket for a compute stack
      required:
      - stack_small_id
      - selected_node_id
      - num_claimed_compute_units
      - requested_attestation_nodes
      - committed_stack_proofs
      - stack_merkle_leaves
      - already_attested_nodes
      - is_in_dispute
      - user_refund_amount
      - is_claimed
      properties:
        already_attested_nodes:
          type: string
          description: Comma-separated list of node IDs that have already attested
        committed_stack_proofs:
          type: array
          items:
            type: integer
            format: int32
            minimum: 0
          description: Cryptographic proof of the committed stack state
        dispute_settled_at_epoch:
          type:
          - integer
          - 'null'
          format: int64
          description: Optional epoch timestamp when a dispute was settled
        is_claimed:
          type: boolean
          description: Indicates whether the settlement ticket has been claimed
        is_in_dispute:
          type: boolean
          description: Indicates whether the stack is currently in a dispute
        num_claimed_compute_units:
          type: integer
          format: int64
          description: Number of compute units claimed to be processed
        requested_attestation_nodes:
          type: string
          description: Comma-separated list of node IDs requested for attestation
        selected_node_id:
          type: integer
          format: int64
          description: Identifier of the node selected for computation
        stack_merkle_leaves:
          type: array
          items:
            type: integer
            format: int32
            minimum: 0
          description: Merkle leaf representing the stack in a larger tree structure
        stack_small_id:
          type: integer
          format: int64
          description: Unique small integer identifier for the stack
        user_refund_amount:
          type: integer
          format: int64
          description: Amount to be refunded to the user (likely in smallest currency unit)
    Task:
      type: object
      description: Represents a task in the system
      required:
      - task_small_id
      - task_id
      - role
      - is_deprecated
      - security_level
      properties:
        deprecated_at_epoch:
          type:
          - integer
          - 'null'
          format: int64
          description: Optional epoch timestamp when the task was deprecated
        is_deprecated:
          type: boolean
          description: Indicates whether the task is deprecated
        minimum_reputation_score:
          type:
          - integer
          - 'null'
          format: int64
          description: Optional minimum reputation score required for the task
        model_name:
          type:
          - string
          - 'null'
          description: Optional name of the model used for the task
        role:
          type: integer
          format: int64
          description: Role associated with the task (encoded as an integer)
        security_level:
          type: integer
          format: int64
          description: Security level of the task (encoded as an integer)
        task_id:
          type: string
          description: Unique string identifier for the task
        task_small_id:
          type: integer
          format: int64
          description: Unique small integer identifier for the task
        valid_until_epoch:
          type:
          - integer
          - 'null'
          format: int64
          description: Optional epoch timestamp until which the task is valid
...